backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.35
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
